requirements.txt
# Career Path Recommender System - Requirements
# Multi-agent system using AWS Bedrock Claude Sonnet 4.5

# Web Interface
streamlit>=1.28.0

# AWS and Bedrock
uv
boto3>=1.34.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-aws>=0.0.20

# Web Scraping and Search
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
ddgs>=8.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Database (sqlite3 is built-in Python module)
sqlalchemy>=2.0.0

# HTTP and Networking
aiohttp>=3.8.0
httpx>=0.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
click>=8.0.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: Enhanced scraping
selenium>=4.15.0
scrapy>=2.10.0

# Optional: Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Optional: Enhanced AI capabilities
openai>=1.0.0
anthropic>=0.7.0
